UNet

Documentation for UNet.

UNet.chcat

UNet.uchain

UNet.unet

UNet.chcat — Method

chcat(x...)

Concatenate the image data along the dimension corresponding to the channels. Image data should be stored in WHCN order (width, height, channels, batch) or WHDCN (width, height, depth, channels, batch) in 3D context. Channels are assumed to be the penultimate dimension.

Example

julia> x1 = rand(32, 32, 4, 6); # a batch of 6 images (32x32) with 4 channels

julia> x2 = rand(32, 32, 4, 6); # a batch of 6 images (32x32) with 4 channels

julia> chcat(x1, x2) |> size
(32, 32, 8, 6)

UNet.uchain — Method

uchain(;encoders, decoders, bridge, connection)

Build a Chain with U-Net like architecture. encoders and decoders are arrays of encoding/decoding blocks, from top to bottom (see diagram below). bridge is the bottom part of U-Net architecture. Each level of the U-Net is connected through a 2-argument callable connection. connection could be an array in case the way levels are connected vary from one level to another.

Notes :

usually encoder block starts with a 'MaxPool' to downsample image by 2, except

the first level encoder.

usually decoder block ends with a 'ConvTranspose' to upsample image by 2,

except the first level decoder.

+---------+                                                          +---------+
|encoder 1|                                                          |decoder 1|
+---------+                                                          +---------+
     |------------------------------------------------------------------->^     
     |                                                                    |     
     |   +---------+                                         +---------+  |     
     +-->|encoder 2|                                         |decoder 2|--+     
         +---------+                                         +---------+        
              |-------------------------------------------------->^             
              |                                                   |             
              |   +---------+                        +---------+  |             
              +-->|encoder 3|                        |decoder 3|--+             
                  +---------+                        +---------+                
                       |--------------------------------->^                     
                       |                                  |                     
                       |   +---------+       +---------+  |                     
                       +-->|encoder 4|       |decoder 4|--+                     
                           +---------+       +---------+                        
                                |---------------->^                             
                                |                 |                             
                                |    +--------+   |                             
                                +--->| bridge |---+                             
                                     +--------+                                  

See also chcat.

UNet.unet — Method

unet(; inchannels)

Build a U-Net to process images with inchannels channels (inchannels = 3 for RGB images). This implementation corresponds to original paper with unpadded convolutions.