UNet

Documentation for UNet.

UNet.adjustsize

UNet.chcat

UNet.checksize

UNet.minsize

UNet.outputsize

UNet.padding

UNet.uchain

UNet.unet

UNet.adjustsize — Method

adjustsize(sz; padding, nlevels)

Computes an image size compatible with a gigen U-Net architecture, from a given image size sz (Int or Tuple of Int).

UNet.chcat — Method

chcat(x...)

Concatenate the image data along the dimension corresponding to the channels. Image data should be stored in WHCN order (width, height, channels, batch) or WHDCN (width, height, depth, channels, batch) in 3D context. Channels are assumed to be the penultimate dimension.

Example

julia> x1 = rand(32, 32, 4, 6); # a batch of 6 images (32x32) with 4 channels

julia> x2 = rand(32, 32, 4, 6); # a batch of 6 images (32x32) with 4 channels

julia> chcat(x1, x2) |> size
(32, 32, 8, 6)

UNet.checksize — Method

checksize(sz; padding, nlevels)

Check if an image size is appropriate with a given U-Net architecture. sz could be either an Int or a Tuple of Int.

UNet.minsize — Method

minsize(; padding, nlevels)
minsize(p, n)

Computes the minimal image size for a given U-Net architecture defined by padding and nlevels.

UNet.outputsize — Method

outputsize(sz; padding, nlevels)

Computes output size for a given U-Net architecture defined by padding and nlevels.

UNet.padding — Method

padding(sz; padding, nlevels)

Computes two tuples of padding values for both input image and ground truth image.

sz: size of input/ground truth image

padding: indicates if convolutions are padded or not

nlevels: number of levels in U-Net architecture

Example

julia> using Images

julia> using UNet: padding

julia> img = fill(RGB(1, 1, 1), (256, 256));

julia> gth = fill(RGB(1, 1, 1), (256, 256));

julia> ip, gp = padding(size(img), padding = false, nlevels = 4)
(((94, 94), (94, 94)), ((2, 2), (2, 2)))

julia> pimg = padarray(img, Pad(:reflect, ip...));

julia> pgth = padarray(gth, Pad(:reflect, gp...));

UNet.uchain — Method

uchain(;encoders, decoders, bridge, connection)

Build a Chain with U-Net like architecture. encoders and decoders are arrays of encoding/decoding blocks, from top to bottom (see diagram below). bridge is the bottom part of U-Net architecture. Each level of the U-Net is connected through a 2-argument callable connection. connection could be an array in case the way levels are connected vary from one level to another.

Notes :

usually encoder block starts with a 'MaxPool' to downsample image by 2, except

the first level encoder.

usually decoder block ends with a 'ConvTranspose' to upsample image by 2,

except the first level decoder.

+---------+                                                          +---------+
|encoder 1|                                                          |decoder 1|
+---------+                                                          +---------+
     |------------------------------------------------------------------->^     
     |                                                                    |     
     |   +---------+                                         +---------+  |     
     +-->|encoder 2|                                         |decoder 2|--+     
         +---------+                                         +---------+        
              |-------------------------------------------------->^             
              |                                                   |             
              |   +---------+                        +---------+  |             
              +-->|encoder 3|                        |decoder 3|--+             
                  +---------+                        +---------+                
                       |--------------------------------->^                     
                       |                                  |                     
                       |   +---------+       +---------+  |                     
                       +-->|encoder 4|       |decoder 4|--+                     
                           +---------+       +---------+                        
                                |---------------->^                             
                                |                 |                             
                                |    +--------+   |                             
                                +--->| bridge |---+                             
                                     +--------+                                  

See also chcat.

UNet.unet — Method

unet(; inchannels, nclasses = 1, volume = false, base = 64,
batchnorm = false, padding = false, upsample = :convt, nlevels = 4)

Build a U-Net to process images with inchannels channels (e.g. inchannels = 3 for RGB images). Default argument values correspond to the original paper implementation with unpadded convolutions.

nclasses: number of pixel classes. If nclasses = 2 the final layer have one channel with a sigmoid activation function, otherwise the final layer have nclasses channels followed by a softmax function.

volume: set to true to process tri-dimensional datas

basewidth: base number of channels, the number of filters/channels is multiplied by two at each U-Net level. Notice that basewidth should be a power of two number, other values could lead to unfunctionnal U-Net.

batchnorm: if true add a BatchNorm layer after each convolution

padding: if true convolutions are padded

upsample: either :convt (ConvTranspose layer), :nearest or :bilinear (Upsample layer)

nlevels: number of level or depth of the U-Net